[Twinkle#67] Refactor Direct SQL Calls to jOOQ Queries

+++ twinkle-app/twinkle.main/src/main/java/com/github/donkeyrit/twinkle/panels/Content/CarPanel.java
@@ -16,14 +16,14 @@ public class CarPanel extends JPanel
 {	
 	private String nameCountry;
 	private String status;
-	private int imagesNum;
+	private long imagesNum;
 
 	@Inject
 	public CarPanel(CarService carService, @Assisted Car car) 
 	{
 		setLayout(null);
 
-		int carId = car.getId();
+		long carId = car.getId();
 		this.imagesNum = carId;
 		this.nameCountry = car.getModelOfCar().getMark().getCountry().getCountryName();
 		this.status = carService.isTaken(carId);
